+++ OpenLocoTool/Data/ObjectType.cs
