+++ OpenLocoTool/Data/SourceGame.cs
@@ -1,0 +1,9 @@
+﻿namespace OpenLocoTool.Data
+{
+	public enum SourceGame : byte
+	{
+		Custom = 0,
+		DataFile = 1,
+		Vanilla = 2,
+	}
+}
